Experimental:
While testing changing millitm to uint32_t the ecm requests to my readers start to give random timeouts
But on higher debugging level timeouts disappeared again.
Changing it even further up from uint16_t to uint64_t and the timeouts disappeared completely.

I seem to unexperienced to find the real reason for this behaviour but for now I want to test if this fixes more issues some users reported! 


git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@10129 4b0bc96b-bc66-0410-9d44-ebda105a78c1

+++ globals.h
@@ -514,7 +514,7 @@ enum {E2_GLOBAL = 0, E2_GROUP, E2_CAID, E2_IDENT, E2_CLASS, E2_CHID, E2_QUEUE, E
 struct timeb
 {
 	time_t time;
-	uint16_t millitm;
+	uint64_t millitm;
 };
 
 typedef struct cs_mutexlock
